% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{sj_update}
\alias{sj_update}
\title{Update sj-packages and its dependencies from CRAN, if necessary.}
\usage{
sj_update(which = c("all", "core", "deps"))
}
\arguments{
\item{which}{String, indicates whether sj-packages (\code{which = "core"}), dependencies (\code{which = "deps"}) or both (\code{which = "all"}) should be checked for available updates.}
}
\description{
Update sj-packages and its dependencies from CRAN, if necessary.
}
