\encoding{UTF-8}
\name{strengejacke-package}
\alias{strengejacke-package}
\alias{strengejacke}
\docType{package}
\title{Strenge Jacke!}
\description{
This package simply loads the packages associated with Strenge Jacke!, namely: \cr
sjPlot, sjmisc and sjstats.
}
\details{
\tabular{rl}{
Package: \tab strengejacke\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2016-10-18\cr
License: \tab GPL-3\cr
}
}
\author{
Daniel Lüdecke \email{d.luedecke@uke.de}
}
